Create address package
